AgentCPM-Explore/AgentDock/master/requirements.txt
# AgentDock Manager Dependencies
# ================================

# Web Framework
fastapi>=0.100.0
uvicorn>=0.23.0
gunicorn>=21.0.0

# Async HTTP Client
httpx>=0.24.0

# Docker SDK
docker>=6.1.0

# Database (MongoDB)
motor>=3.3.0
beanie==1.23.6

# Configuration
PyYAML>=6.0

# Utilities
psutil>=5.9.0
tenacity>=8.2.0
python-multipart>=0.0.6

# Templates
jinja2>=3.1.0
